+++ evmd/config/activators.go
@@ -1,4 +1,4 @@
-package evmd
+package config
 
 import (
 	"github.com/ethereum/go-ethereum/core/vm"
