--- evmd/config/client.toml
@@ -1,17 +1,0 @@
-# This is a TOML config file.
-# For more information, see https://github.com/toml-lang/toml
-
-###############################################################################
-###                           Client Configuration                            ###
-###############################################################################
-
-# The network chain ID
-chain-id = ""
-# The keyring's backend, where the keys are stored (os|file|kwallet|pass|test|memory)
-keyring-backend = "os"
-# CLI output format (text|json)
-output = "text"
-# <host>:<port> to CometBFT RPC interface for this chain
-node = "tcp://localhost:26657"
-# Transaction broadcasting mode (sync|async)
-broadcast-mode = "sync"
